+++ receiver/statsdreceiver/protocol/statsd_parser.go
@@ -164,7 +164,6 @@ func (p *StatsDParser) GetMetrics() pmetric.Metrics {
 		)
 	}
 
-	p.lastIntervalTime = timeNowFunc()
 	p.gauges = make(map[statsDMetricDescription]pmetric.ScopeMetrics)
 	p.counters = make(map[statsDMetricDescription]pmetric.ScopeMetrics)
 	p.timersAndDistributions = make([]pmetric.ScopeMetrics, 0)
@@ -209,10 +208,15 @@ func (p *StatsDParser) Aggregate(line string) error {
 	case CounterType:
 		_, ok := p.counters[parsedMetric.description]
 		if !ok {
-			p.counters[parsedMetric.description] = buildCounterMetric(parsedMetric, p.isMonotonicCounter, timeNowFunc(), p.lastIntervalTime)
+			timeNow := timeNowFunc()
+			p.counters[parsedMetric.description] = buildCounterMetric(parsedMetric, p.isMonotonicCounter, timeNow, p.lastIntervalTime)
+			p.lastIntervalTime = timeNow
+
+			fmt.Println("First call")
 		} else {
 			point := p.counters[parsedMetric.description].Metrics().At(0).Sum().DataPoints().At(0)
 			point.SetIntVal(point.IntVal() + parsedMetric.counterValue())
+			fmt.Println("Repeated call in the same aggreation")
 		}
 
 	case TimingType, HistogramType: